move to lila.web.ui.PiecesSetImages and add a live setting - WIP
as we may need to disabled it, if it creates more missing images than it
solves

TODO: disabling doesn't work yet because some of the PR needs reverting
to get the base64 images back
